requirements.txt
pydub
vosk
moviepy
pytubefix
google-genai
python-dotenv
audioop-lts